requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
python-pptx>=0.6.23
# mcp>=1.0.0  # Python 3.10+ 필요, Claude CLI가 별도 실행하므로 Bot에서는 불필요
